Remove confirmation proposal actions (#237)
* fix: removing confirmation step when cancelling, signing off, finalizing proposal.

* fix: adding description to errors and changed console.log to console.error
